jqueryで書かれているものをvanillajsに変更
