+++ app/javascript/application.js
@@ -16,6 +16,7 @@ import "./src/jquery-ui"
 import "./src/direct_uploads"
 import "./src/tier_item_updater"
 import "./src/dynamic_field"
+import "./src/tier_image_uploader"
 
 ActiveStorage.start()
 
@@ -39,38 +40,4 @@ $(function() {
       }
     }
   });
-
-  $(document).on('click', '#open-modal', function() {
-    $("#myModal").dialog("open");
-  });
-
-  $(document).on('change', '#image-upload', function() {
-    let numberOfFiles = $(this)[0].files.length;
-    if (numberOfFiles > 30) {
-      alert('画像は30枚以下にしてください');
-      $(this).val('');
-    }
-  });
-
-  showUploadedImagePreview('#image-upload');
-  showUploadedImagePreview('#cover-image-upload');
-
-  function showUploadedImagePreview(inputSelector) {
-    $(document).on('change', inputSelector, function() {
-      let images = this.files;
-      let previewArea = $(this).next('.image-preview');
-      previewArea.empty();
-  
-      $.each(images, function(index, image) {
-        let reader = new FileReader();
-  
-        reader.onload = function(e) {
-          let img = $('<img>').attr('src', e.target.result).css({width: '80px', height: 'auto'});
-          previewArea.append(img);
-        };
-  
-        reader.readAsDataURL(image);
-      });
-    });
-  }
 });
